restructured test directory
